ls_6 Sliders calling from index.js
